+++ src/src/Stepper.cpp
@@ -39,8 +39,8 @@ volatile int64_t calculated_stepper_pulses=0;
 
 volatile bool jog_done = true;
 volatile int32_t jog_steps = 0;
-float jog_mm = 0;
-volatile float jog_scaler = 0.2;
+double jog_mm = 0;
+volatile double jog_scaler = 0.2;
 volatile uint16_t vel = 1;
 volatile double stopPos = 0;
 volatile double stopNeg = 0;
